refactor: split image_video_dataset.py into focused modules (2417 -> 901 lines)
Extract five modules from the monolithic dataset file for better readability:
- architectures.py: ARCHITECTURE_* constants
- media_utils.py: image/video file I/O utilities (glob, load, resize)
- cache_io.py: latent/text encoder cache save functions (all architectures)
- bucket.py: BucketSelector and BucketBatchManager
- datasources.py: Content/Image/Video Datasource class hierarchy

Backward compatibility maintained via re-exports in image_video_dataset.py.
