+++ src/musubi_tuner/dataset/image_video_dataset.py
@@ -1,7 +1,5 @@
 from concurrent.futures import ThreadPoolExecutor
 import glob
-from importlib.util import find_spec
-import json
 import math
 import os
 import random
@@ -16,165 +14,18 @@ SharedEpoch = Optional["Synchronized[int]"]
 
 import numpy as np
 import torch
-from safetensors.torch import save_file, load_file
 from PIL import Image
-import cv2
-import av
 
 from musubi_tuner.utils import safetensors_utils
-from musubi_tuner.utils.model_utils import dtype_to_str
 
 import logging
 
 logger = logging.getLogger(__name__)
 logging.basicConfig(level=logging.INFO)
 
-
-IMAGE_EXTENSIONS = [".png", ".jpg", ".jpeg", ".webp", ".bmp", ".PNG", ".JPG", ".JPEG", ".WEBP", ".BMP", ".avif", ".AVIF"]
-
-
-if find_spec("jxlpy") is not None:  # JPEG-XL on Linux
-    from jxlpy import JXLImagePlugin  # noqa: F401 # type: ignore
-
-    IMAGE_EXTENSIONS.extend([".jxl", ".JXL"])
-
-if find_spec("pillow_jxl") is not None:  # JPEG-XL on Windows
-    import pillow_jxl  # noqa: F401 # type: ignore
-
-    IMAGE_EXTENSIONS.extend([".jxl", ".JXL"])
-
-VIDEO_EXTENSIONS = [
-    ".mp4",
-    ".webm",
-    ".avi",
-    ".mkv",
-    ".mov",
-    ".flv",
-    ".wmv",
-    ".m4v",
-    ".mpg",
-    ".mpeg",
-    ".MP4",
-    ".WEBM",
-    ".AVI",
-    ".MKV",
-    ".MOV",
-    ".FLV",
-    ".WMV",
-    ".M4V",
-    ".MPG",
-    ".MPEG",
-]  # some of them are not tested
-
-# Architecture short names cannot contain underscore
-ARCHITECTURE_HUNYUAN_VIDEO = "hv"
-ARCHITECTURE_HUNYUAN_VIDEO_FULL = "hunyuan_video"
-ARCHITECTURE_WAN = "wan"
-ARCHITECTURE_WAN_FULL = "wan"
-ARCHITECTURE_FRAMEPACK = "fp"
-ARCHITECTURE_FRAMEPACK_FULL = "framepack"
-ARCHITECTURE_FLUX_KONTEXT = "fk"
-ARCHITECTURE_FLUX_KONTEXT_FULL = "flux_kontext"
-ARCHITECTURE_FLUX_2_DEV = "f2d"
-ARCHITECTURE_FLUX_2_DEV_FULL = "flux_2_dev"
-ARCHITECTURE_FLUX_2_KLEIN_4B = "f2k4b"
-ARCHITECTURE_FLUX_2_KLEIN_4B_FULL = "flux_2_klein_4b"
-ARCHITECTURE_FLUX_2_KLEIN_9B = "f2k9b"
-ARCHITECTURE_FLUX_2_KLEIN_9B_FULL = "flux_2_klein_9b"
-ARCHITECTURE_QWEN_IMAGE = "qi"
-ARCHITECTURE_QWEN_IMAGE_FULL = "qwen_image"
-ARCHITECTURE_QWEN_IMAGE_EDIT = "qie"
-ARCHITECTURE_QWEN_IMAGE_EDIT_FULL = "qwen_image_edit"
-ARCHITECTURE_QWEN_IMAGE_LAYERED = "qil"
-ARCHITECTURE_QWEN_IMAGE_LAYERED_FULL = "qwen_image_layered"
-ARCHITECTURE_KANDINSKY5 = "k5"
-ARCHITECTURE_KANDINSKY5_FULL = "kandinsky5"
-ARCHITECTURE_HUNYUAN_VIDEO_1_5 = "hv15"
-ARCHITECTURE_HUNYUAN_VIDEO_1_5_FULL = "hunyuan_video_1_5"
-ARCHITECTURE_Z_IMAGE = "zi"
-ARCHITECTURE_Z_IMAGE_FULL = "z_image"
-
-
-def glob_images(directory, base="*", caption_extension=None):
-    img_paths = []
-    for ext in IMAGE_EXTENSIONS:
-        if base == "*":
-            img_paths.extend(glob.glob(os.path.join(glob.escape(directory), base + ext)))
-        else:
-            img_paths.extend(glob.glob(glob.escape(os.path.join(directory, base + ext))))
-    img_paths = list(set(img_paths))  # remove duplicates
-
-    # check for caption files and only keep images with captions
-    if caption_extension is not None:
-        caption_paths = glob.glob(os.path.join(glob.escape(directory), "*" + caption_extension))
-        caption_bases = set()
-        for caption_path in caption_paths:
-            caption_base = os.path.splitext(os.path.basename(caption_path))[0]
-            caption_bases.add(caption_base)
-        filtered_img_paths = []
-        for img_path in img_paths:
-            img_base = os.path.splitext(os.path.basename(img_path))[0]
-            if img_base in caption_bases:
-                filtered_img_paths.append(img_path)
-        img_paths = filtered_img_paths
-
-    img_paths.sort()
-    return img_paths
-
-
-def glob_videos(directory, base="*"):
-    video_paths = []
-    for ext in VIDEO_EXTENSIONS:
-        if base == "*":
-            video_paths.extend(glob.glob(os.path.join(glob.escape(directory), base + ext)))
-        else:
-            video_paths.extend(glob.glob(glob.escape(os.path.join(directory, base + ext))))
-    video_paths = list(set(video_paths))  # remove duplicates
-    video_paths.sort()
-    return video_paths
-
-
-def divisible_by(num: int, divisor: int) -> int:
-    return num - num % divisor
-
-
-def resize_image_to_bucket(image: Union[Image.Image, np.ndarray], bucket_reso: tuple[int, int]) -> np.ndarray:
-    """
-    Resize the image to the bucket resolution.
-
-    bucket_reso: **(width, height)**
-    """
-    is_pil_image = isinstance(image, Image.Image)
-    if is_pil_image:
-        image_width, image_height = image.size
-    else:
-        image_height, image_width = image.shape[:2]
-
-    if bucket_reso == (image_width, image_height):
-        return np.array(image) if is_pil_image else image
-
-    bucket_width, bucket_height = bucket_reso
-
-    # resize the image to the bucket resolution to match the short side
-    scale_width = bucket_width / image_width
-    scale_height = bucket_height / image_height
-    scale = max(scale_width, scale_height)
-    image_width = int(image_width * scale + 0.5)
-    image_height = int(image_height * scale + 0.5)
-
-    if scale > 1:
-        image = Image.fromarray(image) if not is_pil_image else image
-        image = image.resize((image_width, image_height), Image.LANCZOS)
-        image = np.array(image)
-    else:
-        image = np.array(image) if is_pil_image else image
-        image = cv2.resize(image, (image_width, image_height), interpolation=cv2.INTER_AREA)
-
-    # crop the image to the bucket resolution
-    crop_left = (image_width - bucket_width) // 2
-    crop_top = (image_height - bucket_height) // 2
-    image = image[crop_top : crop_top + bucket_height, crop_left : crop_left + bucket_width]
-    return image
+from musubi_tuner.dataset.architectures import *  # noqa: F401,F403
+from musubi_tuner.dataset.media_utils import *  # noqa: F401,F403
+from musubi_tuner.dataset.media_utils import glob_images, glob_videos, resize_image_to_bucket, load_video  # explicit imports for local use
 
 
 class ItemInfo:
@@ -216,1394 +67,27 @@ class ItemInfo:
         )
 
 
-# We use simple if-else approach to support multiple architectures.
-# Maybe we can use a plugin system in the future.
+from musubi_tuner.dataset.cache_io import *  # noqa: F401,F403
 
-# the keys of the dict are `<content_type>_FxHxW_<dtype>` for latents
-# and `<content_type>_<dtype|mask>` for other tensors
 
+from musubi_tuner.dataset.bucket import BucketSelector, BucketBatchManager  # noqa: F401
 
-def save_latent_cache(item_info: ItemInfo, latent: torch.Tensor):
-    """HunyuanVideo architecture. HunyuanVideo doesn't support I2V and control latents"""
-    assert latent.dim() == 4, "latent should be 4D tensor (frame, channel, height, width)"
 
-    _, F, H, W = latent.shape
-    dtype_str = dtype_to_str(latent.dtype)
-    sd = {f"latents_{F}x{H}x{W}_{dtype_str}": latent.detach().cpu()}
+from musubi_tuner.dataset.datasources import (  # noqa: F401
+    ContentDatasource,
+    ImageDatasource,
+    ImageDirectoryDatasource,
+    ImageJsonlDatasource,
+    VideoDatasource,
+    VideoDirectoryDatasource,
+    VideoJsonlDatasource,
+)
 
-    save_latent_cache_common(item_info, sd, ARCHITECTURE_HUNYUAN_VIDEO_FULL)
 
-
-def save_latent_cache_wan(
-    item_info: ItemInfo,
-    latent: torch.Tensor,
-    clip_embed: Optional[torch.Tensor],
-    image_latent: Optional[torch.Tensor],
-    control_latent: Optional[torch.Tensor],
-    f_indices: Optional[list[int]] = None,
-):
-    """Wan architecture"""
-    assert latent.dim() == 4, "latent should be 4D tensor (frame, channel, height, width)"
-
-    _, F, H, W = latent.shape
-    dtype_str = dtype_to_str(latent.dtype)
-    sd = {f"latents_{F}x{H}x{W}_{dtype_str}": latent.detach().cpu()}
-
-    if clip_embed is not None:
-        sd[f"clip_{dtype_str}"] = clip_embed.detach().cpu()
-
-    if image_latent is not None:
-        sd[f"latents_image_{F}x{H}x{W}_{dtype_str}"] = image_latent.detach().cpu()
-
-    if control_latent is not None:
-        sd[f"latents_control_{F}x{H}x{W}_{dtype_str}"] = control_latent.detach().cpu()
-
-    if f_indices is not None:
-        dtype_str = dtype_to_str(torch.int32)
-        sd[f"f_indices_{dtype_str}"] = torch.tensor(f_indices, dtype=torch.int32)
-
-    save_latent_cache_common(item_info, sd, ARCHITECTURE_WAN_FULL)
-
-
-def save_latent_cache_framepack(
-    item_info: ItemInfo,
-    latent: torch.Tensor,
-    latent_indices: torch.Tensor,
-    clean_latents: torch.Tensor,
-    clean_latent_indices: torch.Tensor,
-    clean_latents_2x: torch.Tensor,
-    clean_latent_2x_indices: torch.Tensor,
-    clean_latents_4x: torch.Tensor,
-    clean_latent_4x_indices: torch.Tensor,
-    image_embeddings: torch.Tensor,
-):
-    """FramePack architecture"""
-    assert latent.dim() == 4, "latent should be 4D tensor (frame, channel, height, width)"
-
-    _, F, H, W = latent.shape
-    dtype_str = dtype_to_str(latent.dtype)
-    sd = {f"latents_{F}x{H}x{W}_{dtype_str}": latent.detach().cpu().contiguous()}
-
-    # `latents_xxx` must have {F, H, W} suffix
-    indices_dtype_str = dtype_to_str(latent_indices.dtype)
-    sd[f"image_embeddings_{dtype_str}"] = image_embeddings.detach().cpu()  # image embeddings dtype is same as latents dtype
-    sd[f"latent_indices_{indices_dtype_str}"] = latent_indices.detach().cpu()
-    sd[f"clean_latent_indices_{indices_dtype_str}"] = clean_latent_indices.detach().cpu()
-    sd[f"latents_clean_{F}x{H}x{W}_{dtype_str}"] = clean_latents.detach().cpu().contiguous()
-    if clean_latent_2x_indices is not None:
-        sd[f"clean_latent_2x_indices_{indices_dtype_str}"] = clean_latent_2x_indices.detach().cpu()
-    if clean_latents_2x is not None:
-        sd[f"latents_clean_2x_{F}x{H}x{W}_{dtype_str}"] = clean_latents_2x.detach().cpu().contiguous()
-    if clean_latent_4x_indices is not None:
-        sd[f"clean_latent_4x_indices_{indices_dtype_str}"] = clean_latent_4x_indices.detach().cpu()
-    if clean_latents_4x is not None:
-        sd[f"latents_clean_4x_{F}x{H}x{W}_{dtype_str}"] = clean_latents_4x.detach().cpu().contiguous()
-
-    # for key, value in sd.items():
-    #     print(f"{key}: {value.shape}")
-    save_latent_cache_common(item_info, sd, ARCHITECTURE_FRAMEPACK_FULL)
-
-
-def save_latent_cache_flux_kontext(
-    item_info: ItemInfo,
-    latent: torch.Tensor,
-    control_latent: torch.Tensor,
-):
-    """FLUX.1 Kontext architecture"""
-    assert latent.dim() == 3, "latent should be 3D tensor (channel, height, width)"
-
-    _, H, W = latent.shape
-    F = 1
-    dtype_str = dtype_to_str(latent.dtype)
-    sd = {f"latents_{F}x{H}x{W}_{dtype_str}": latent.detach().cpu().contiguous()}
-
-    _, H, W = control_latent.shape
-    F = 1
-    sd[f"latents_control_{F}x{H}x{W}_{dtype_str}"] = control_latent.detach().cpu().contiguous()
-
-    save_latent_cache_common(item_info, sd, ARCHITECTURE_FLUX_KONTEXT_FULL)
-
-
-def save_latent_cache_flux_2(
-    item_info: ItemInfo, latent: torch.Tensor, control_latent: Optional[list[torch.Tensor]], arch_full: str
-):
-    """Flux 2 architecture"""
-    assert latent.dim() == 3, "latent should be 3D tensor (channel, height, width)"
-    assert control_latent is None or all(cl.dim() == 3 for cl in control_latent), (
-        "control_latent should be 3D tensor (channel, height, width) or None"
-    )
-
-    _, H, W = latent.shape
-    dtype_str = dtype_to_str(latent.dtype)
-    sd = {f"latents_{H}x{W}_{dtype_str}": latent.detach().cpu().contiguous()}
-
-    if control_latent is not None:
-        for i, cl in enumerate(control_latent):
-            _, H, W = cl.shape
-            sd[f"latents_control_{i}_{H}x{W}_{dtype_str}"] = cl.detach().cpu().contiguous()
-
-    save_latent_cache_common(item_info, sd, arch_full)
-
-
-def save_latent_cache_qwen_image(item_info: ItemInfo, latent: torch.Tensor, control_latent: Optional[list[torch.Tensor]]):
-    """Qwen-Image architecture"""
-    assert latent.dim() == 4, "latent should be 4D tensor (frame, channel, height, width)"
-    assert control_latent is None or all(cl.dim() == 4 for cl in control_latent), (
-        "control_latent should be 4D tensor (frame, channel, height, width) or None"
-    )
-
-    _, F, H, W = latent.shape
-    dtype_str = dtype_to_str(latent.dtype)
-    sd = {f"latents_{F}x{H}x{W}_{dtype_str}": latent.detach().cpu().contiguous()}
-
-    if control_latent is not None:
-        for i, cl in enumerate(control_latent):
-            _, F, H, W = cl.shape
-            sd[f"latents_control_{i}_{F}x{H}x{W}_{dtype_str}"] = cl.detach().cpu().contiguous()
-
-    save_latent_cache_common(item_info, sd, ARCHITECTURE_QWEN_IMAGE_FULL)
-
-
-def save_latent_cache_kandinsky5(
-    item_info: ItemInfo,
-    latent: torch.Tensor,
-    image_latent: Optional[torch.Tensor] = None,
-    control_latent: Optional[torch.Tensor] = None,
-    scaling_factor: Optional[float] = None,
-):
-    """Kandinsky 5 architecture (image/video), with optional source/control latents for i2v/control."""
-    assert latent.dim() == 3 or latent.dim() == 4, "latent should be 3D (C,H,W) or 4D (F,C,H,W) tensor"
-
-    if latent.dim() == 4:
-        _, F, H, W = latent.shape
-    else:
-        F, H, W = 1, latent.shape[1], latent.shape[2]
-        latent = latent.unsqueeze(0)
-    dtype_str = dtype_to_str(latent.dtype)
-    sd = {f"latents_{F}x{H}x{W}_{dtype_str}": latent.detach().cpu().contiguous().clone()}
-
-    if image_latent is not None:
-        _, F_img, H_img, W_img = image_latent.shape
-        sd[f"latents_image_{F_img}x{H_img}x{W_img}_{dtype_str}"] = image_latent.detach().cpu().contiguous().clone()
-
-    if control_latent is not None:
-        _, F_ctrl, H_ctrl, W_ctrl = control_latent.shape
-        sd[f"latents_control_{F_ctrl}x{H_ctrl}x{W_ctrl}_{dtype_str}"] = control_latent.detach().cpu().contiguous().clone()
-
-    if scaling_factor is not None:
-        sd["vae_scaling_factor"] = torch.tensor(float(scaling_factor))
-
-    save_latent_cache_common(item_info, sd, ARCHITECTURE_KANDINSKY5_FULL)
-
-
-def save_latent_cache_hunyuan_video_1_5(
-    item_info: ItemInfo,
-    latent: torch.Tensor,
-    image_latent: Optional[torch.Tensor],
-    vision_feature: Optional[torch.Tensor],
-):
-    """HunyuanVideo 1.5 architecture"""
-    _, F, H, W = latent.shape
-    dtype_str = dtype_to_str(latent.dtype)
-    sd: dict[str, torch.Tensor] = {f"latents_{F}x{H}x{W}_{dtype_str}": latent.detach().cpu()}
-
-    if image_latent is not None:
-        dtype_str = dtype_to_str(image_latent.dtype)
-        _, F, H, W = image_latent.shape
-        sd[f"latents_image_{F}x{H}x{W}_{dtype_str}"] = image_latent.detach().cpu()
-
-    if vision_feature is not None:
-        dtype_str = dtype_to_str(vision_feature.dtype)
-        sd[f"siglip_{dtype_str}"] = vision_feature.detach().cpu()
-
-    save_latent_cache_common(item_info, sd, ARCHITECTURE_HUNYUAN_VIDEO_1_5_FULL)
-
-
-def save_latent_cache_z_image(item_info: ItemInfo, latent: torch.Tensor):
-    """Z-Image architecture. No control latent is supported."""
-    assert latent.dim() == 3, "latent should be 3D tensor (channel, height, width)"
-
-    C, H, W = latent.shape
-    F = 1
-    dtype_str = dtype_to_str(latent.dtype)
-    sd = {f"latents_{F}x{H}x{W}_{dtype_str}": latent.detach().cpu().contiguous()}
-
-    save_latent_cache_common(item_info, sd, ARCHITECTURE_Z_IMAGE_FULL)
-
-
-def save_latent_cache_common(item_info: ItemInfo, sd: dict[str, torch.Tensor], arch_fullname: str):
-    metadata = {
-        "architecture": arch_fullname,
-        "width": f"{item_info.original_size[0]}",
-        "height": f"{item_info.original_size[1]}",
-        "format_version": "1.0.1",
-    }
-    if item_info.frame_count is not None:
-        metadata["frame_count"] = f"{item_info.frame_count}"
-
-    for key, value in sd.items():
-        # NaN check and show warning, replace NaN with 0
-        if torch.isnan(value).any():
-            logger.warning(f"{key} tensor has NaN: {item_info.item_key}, replace NaN with 0")
-            value[torch.isnan(value)] = 0
-
-    latent_dir = os.path.dirname(item_info.latent_cache_path)
-    os.makedirs(latent_dir, exist_ok=True)
-
-    save_file(sd, item_info.latent_cache_path, metadata=metadata)
-
-
-def save_text_encoder_output_cache(item_info: ItemInfo, embed: torch.Tensor, mask: Optional[torch.Tensor], is_llm: bool):
-    """HunyuanVideo architecture"""
-    assert embed.dim() == 1 or embed.dim() == 2, (
-        f"embed should be 2D tensor (feature, hidden_size) or (hidden_size,), got {embed.shape}"
-    )
-    assert mask is None or mask.dim() == 1, f"mask should be 1D tensor (feature), got {mask.shape}"
-
-    sd = {}
-    dtype_str = dtype_to_str(embed.dtype)
-    text_encoder_type = "llm" if is_llm else "clipL"
-    sd[f"{text_encoder_type}_{dtype_str}"] = embed.detach().cpu()
-    if mask is not None:
-        sd[f"{text_encoder_type}_mask"] = mask.detach().cpu()
-
-    save_text_encoder_output_cache_common(item_info, sd, ARCHITECTURE_HUNYUAN_VIDEO_FULL)
-
-
-def save_text_encoder_output_cache_wan(item_info: ItemInfo, embed: torch.Tensor):
-    """Wan architecture. Wan2.1 only has a single text encoder"""
-
-    sd = {}
-    dtype_str = dtype_to_str(embed.dtype)
-    text_encoder_type = "t5"
-    sd[f"varlen_{text_encoder_type}_{dtype_str}"] = embed.detach().cpu()
-
-    save_text_encoder_output_cache_common(item_info, sd, ARCHITECTURE_WAN_FULL)
-
-
-def save_text_encoder_output_cache_framepack(
-    item_info: ItemInfo, llama_vec: torch.Tensor, llama_attention_mask: torch.Tensor, clip_l_pooler: torch.Tensor
-):
-    """FramePack architecture."""
-    sd = {}
-    dtype_str = dtype_to_str(llama_vec.dtype)
-    sd[f"llama_vec_{dtype_str}"] = llama_vec.detach().cpu()
-    sd["llama_attention_mask"] = llama_attention_mask.detach().cpu()
-    dtype_str = dtype_to_str(clip_l_pooler.dtype)
-    sd[f"clip_l_pooler_{dtype_str}"] = clip_l_pooler.detach().cpu()
-
-    save_text_encoder_output_cache_common(item_info, sd, ARCHITECTURE_FRAMEPACK_FULL)
-
-
-def save_text_encoder_output_cache_flux_kontext(item_info: ItemInfo, t5_vec: torch.Tensor, clip_l_pooler: torch.Tensor):
-    """Flux Kontext architecture."""
-
-    sd = {}
-    dtype_str = dtype_to_str(t5_vec.dtype)
-    sd[f"t5_vec_{dtype_str}"] = t5_vec.detach().cpu()
-    dtype_str = dtype_to_str(clip_l_pooler.dtype)
-    sd[f"clip_l_pooler_{dtype_str}"] = clip_l_pooler.detach().cpu()
-
-    save_text_encoder_output_cache_common(item_info, sd, ARCHITECTURE_FLUX_KONTEXT_FULL)
-
-
-def save_text_encoder_output_cache_flux_2(item_info: ItemInfo, ctx_vec: torch.Tensor, arch_full: str):
-    """Flux 2 architecture."""
-
-    sd = {}
-    dtype_str = dtype_to_str(ctx_vec.dtype)
-    sd[f"ctx_vec_{dtype_str}"] = ctx_vec.detach().cpu()
-
-    save_text_encoder_output_cache_common(item_info, sd, arch_full)
-
-
-def save_text_encoder_output_cache_qwen_image(item_info: ItemInfo, embed: torch.Tensor):
-    """Qwen-Image architecture."""
-    sd = {}
-    dtype_str = dtype_to_str(embed.dtype)
-    sd[f"varlen_vl_embed_{dtype_str}"] = embed.detach().cpu()
-
-    save_text_encoder_output_cache_common(item_info, sd, ARCHITECTURE_QWEN_IMAGE_FULL)
-
-
-def save_text_encoder_output_cache_kandinsky5(
-    item_info: ItemInfo, text_embeds: torch.Tensor, pooled_embed: torch.Tensor, attention_mask: torch.Tensor
-):
-    """Kandinsky 5 architecture."""
-    sd = {}
-    dtype_str = dtype_to_str(text_embeds.dtype)
-    sd[f"text_embeds_{dtype_str}"] = text_embeds.detach().cpu()
-    dtype_str = dtype_to_str(pooled_embed.dtype)
-    sd[f"pooled_embed_{dtype_str}"] = pooled_embed.detach().cpu()
-    sd["attention_mask"] = attention_mask.detach().cpu()
-
-    save_text_encoder_output_cache_common(item_info, sd, ARCHITECTURE_KANDINSKY5_FULL)
-
-
-def save_text_encoder_output_cache_hunyuan_video_1_5(item_info: ItemInfo, embed: torch.Tensor, byt5_embed: torch.Tensor):
-    """Hunyuan-Video 1.5 architecture."""
-    sd = {}
-    dtype_str = dtype_to_str(embed.dtype)
-    sd[f"varlen_vl_embed_{dtype_str}"] = embed.detach().cpu()
-    dtype_str = dtype_to_str(byt5_embed.dtype)
-    sd[f"varlen_byt5_embed_{dtype_str}"] = byt5_embed.detach().cpu()
-    save_text_encoder_output_cache_common(item_info, sd, ARCHITECTURE_HUNYUAN_VIDEO_1_5_FULL)
-
-
-def save_text_encoder_output_cache_z_image(item_info: ItemInfo, embed: torch.Tensor):
-    """Z-Image architecture."""
-    sd = {}
-    dtype_str = dtype_to_str(embed.dtype)
-    sd[f"varlen_llm_embed_{dtype_str}"] = embed.detach().cpu()
-
-    save_text_encoder_output_cache_common(item_info, sd, ARCHITECTURE_Z_IMAGE_FULL)
-
-
-def save_text_encoder_output_cache_common(item_info: ItemInfo, sd: dict[str, torch.Tensor], arch_fullname: str):
-    for key, value in sd.items():
-        # NaN check and show warning, replace NaN with 0
-        if torch.isnan(value).any():
-            logger.warning(f"{key} tensor has NaN: {item_info.item_key}, replace NaN with 0")
-            value[torch.isnan(value)] = 0
-
-    metadata = {
-        "architecture": arch_fullname,
-        "caption1": item_info.caption,
-        "format_version": "1.0.1",
-    }
-
-    if os.path.exists(item_info.text_encoder_output_cache_path):
-        # load existing cache and update metadata
-        with safetensors_utils.MemoryEfficientSafeOpen(item_info.text_encoder_output_cache_path) as f:
-            existing_metadata = f.metadata()
-            for key in f.keys():
-                if key not in sd:  # avoid overwriting by existing cache, we keep the new one
-                    sd[key] = f.get_tensor(key)
-
-        assert existing_metadata["architecture"] == metadata["architecture"], "architecture mismatch"
-        if existing_metadata["caption1"] != metadata["caption1"]:
-            logger.warning(f"caption mismatch: existing={existing_metadata['caption1']}, new={metadata['caption1']}, overwrite")
-        # TODO verify format_version
-
-        existing_metadata.pop("caption1", None)
-        existing_metadata.pop("format_version", None)
-        metadata.update(existing_metadata)  # copy existing metadata except caption and format_version
-    else:
-        text_encoder_output_dir = os.path.dirname(item_info.text_encoder_output_cache_path)
-        os.makedirs(text_encoder_output_dir, exist_ok=True)
-
-    safetensors_utils.mem_eff_save_file(sd, item_info.text_encoder_output_cache_path, metadata=metadata)
-
-
-class BucketSelector:
-    RESOLUTION_STEPS_HUNYUAN = 16
-    RESOLUTION_STEPS_WAN = 16
-    RESOLUTION_STEPS_FRAMEPACK = 16
-    RESOLUTION_STEPS_FLUX_KONTEXT = 16
-    RESOLUTION_STEPS_FLUX_2 = 16
-    RESOLUTION_STEPS_QWEN_IMAGE = 16
-    RESOLUTION_STEPS_QWEN_IMAGE_EDIT = 16
-    RESOLUTION_STEPS_KANDINSKY5 = 16
-    RESOLUTION_STEPS_HUNYUAN_VIDEO_1_5 = 16
-    RESOLUTION_STEPS_Z_IMAGE = 16
-
-    ARCHITECTURE_STEPS_MAP = {
-        ARCHITECTURE_HUNYUAN_VIDEO: RESOLUTION_STEPS_HUNYUAN,
-        ARCHITECTURE_WAN: RESOLUTION_STEPS_WAN,
-        ARCHITECTURE_FRAMEPACK: RESOLUTION_STEPS_FRAMEPACK,
-        ARCHITECTURE_FLUX_KONTEXT: RESOLUTION_STEPS_FLUX_KONTEXT,
-        ARCHITECTURE_FLUX_2_DEV: RESOLUTION_STEPS_FLUX_2,
-        ARCHITECTURE_FLUX_2_KLEIN_4B: RESOLUTION_STEPS_FLUX_2,
-        ARCHITECTURE_FLUX_2_KLEIN_9B: RESOLUTION_STEPS_FLUX_2,
-        ARCHITECTURE_QWEN_IMAGE: RESOLUTION_STEPS_QWEN_IMAGE,
-        ARCHITECTURE_QWEN_IMAGE_EDIT: RESOLUTION_STEPS_QWEN_IMAGE_EDIT,
-        ARCHITECTURE_QWEN_IMAGE_LAYERED: RESOLUTION_STEPS_QWEN_IMAGE,  # use same steps as Qwen-Image
-        ARCHITECTURE_KANDINSKY5: RESOLUTION_STEPS_KANDINSKY5,
-        ARCHITECTURE_HUNYUAN_VIDEO_1_5: RESOLUTION_STEPS_HUNYUAN_VIDEO_1_5,
-        ARCHITECTURE_Z_IMAGE: RESOLUTION_STEPS_Z_IMAGE,
-    }
-
-    def __init__(
-        self, resolution: Tuple[int, int], enable_bucket: bool = True, no_upscale: bool = False, architecture: str = "no_default"
-    ):
-        self.resolution = resolution
-        self.bucket_area = resolution[0] * resolution[1]
-        self.architecture = architecture
-
-        if architecture in BucketSelector.ARCHITECTURE_STEPS_MAP:
-            self.reso_steps = BucketSelector.ARCHITECTURE_STEPS_MAP[architecture]
-        else:
-            raise ValueError(f"Invalid architecture: {architecture}")
-
-        if not enable_bucket:
-            # only define one bucket
-            self.bucket_resolutions = [resolution]
-            self.no_upscale = False
-        else:
-            # prepare bucket resolution
-            self.no_upscale = no_upscale
-            sqrt_size = int(math.sqrt(self.bucket_area))
-            min_size = divisible_by(sqrt_size // 2, self.reso_steps)
-            self.bucket_resolutions = []
-            for w in range(min_size, sqrt_size + self.reso_steps, self.reso_steps):
-                h = divisible_by(self.bucket_area // w, self.reso_steps)
-                self.bucket_resolutions.append((w, h))
-                self.bucket_resolutions.append((h, w))
-
-            self.bucket_resolutions = list(set(self.bucket_resolutions))
-            self.bucket_resolutions.sort()
-
-        # calculate aspect ratio to find the nearest resolution
-        self.aspect_ratios = np.array([w / h for w, h in self.bucket_resolutions])
-
-    def get_bucket_resolution(self, image_size: tuple[int, int]) -> tuple[int, int]:
-        """
-        return the bucket resolution for the given image size, (width, height)
-        """
-        area = image_size[0] * image_size[1]
-        if self.no_upscale and area <= self.bucket_area:
-            w, h = image_size
-            w = divisible_by(w, self.reso_steps)
-            h = divisible_by(h, self.reso_steps)
-            return w, h
-
-        aspect_ratio = image_size[0] / image_size[1]
-        ar_errors = self.aspect_ratios - aspect_ratio
-        bucket_id = np.abs(ar_errors).argmin()
-        return self.bucket_resolutions[bucket_id]
-
-    @classmethod
-    def calculate_bucket_resolution(
-        cls,
-        image_size: tuple[int, int],
-        resolution: tuple[int, int],
-        reso_steps: Optional[int] = None,
-        architecture: Optional[str] = None,
-    ) -> tuple[int, int]:
-        """
-        Get the bucket resolution for the given image size, resolution and resolution steps.
-        Return (width, height).
-        """
-        if reso_steps is None and architecture is None:
-            raise ValueError("resolution steps or architecture must be provided")
-        if reso_steps is None and architecture is not None:
-            if architecture in BucketSelector.ARCHITECTURE_STEPS_MAP:
-                reso_steps = BucketSelector.ARCHITECTURE_STEPS_MAP[architecture]
-            else:
-                raise ValueError(f"Invalid architecture: {architecture}")
-
-        max_area = resolution[0] * resolution[1]
-        width, height = image_size
-        aspect_ratio = width / height
-        bucket_width = int(math.sqrt(max_area * aspect_ratio))
-        bucket_height = int(math.sqrt(max_area / aspect_ratio))
-        bucket_width = divisible_by(bucket_width, reso_steps)
-        bucket_height = divisible_by(bucket_height, reso_steps)
-
-        # find appropriate resolutions
-        best_resolution = None
-        best_aspect_ratio_diff = float("inf")
-        for i in range(-2, 3):
-            w = bucket_width + i * reso_steps
-            h = divisible_by(max_area // w, reso_steps)
-            current_aspect_ratio_diff = abs((w / h) - aspect_ratio)
-            if current_aspect_ratio_diff < best_aspect_ratio_diff:
-                best_aspect_ratio_diff = current_aspect_ratio_diff
-                best_resolution = (w, h)
-
-        if best_resolution is not None:
-            return best_resolution
-
-        return bucket_width, bucket_height
-
-
-def load_video(
-    video_path: str,
-    start_frame: Optional[int] = None,
-    end_frame: Optional[int] = None,
-    bucket_selector: Optional[BucketSelector] = None,
-    bucket_reso: Optional[tuple[int, int]] = None,
-    source_fps: Optional[float] = None,
-    target_fps: Optional[float] = None,
-) -> list[np.ndarray]:
-    """
-    bucket_reso: if given, resize the video to the bucket resolution, (width, height)
-    """
-    if source_fps is None or target_fps is None:
-        if os.path.isfile(video_path):
-            container = av.open(video_path)
-            video = []
-            for i, frame in enumerate(container.decode(video=0)):
-                if start_frame is not None and i < start_frame:
-                    continue
-                if end_frame is not None and i >= end_frame:
-                    break
-                frame = frame.to_image()
-
-                if bucket_selector is not None and bucket_reso is None:
-                    bucket_reso = bucket_selector.get_bucket_resolution(frame.size)  # calc resolution from first frame
-
-                if bucket_reso is not None:
-                    frame = resize_image_to_bucket(frame, bucket_reso)
-                else:
-                    frame = np.array(frame)
-
-                video.append(frame)
-            container.close()
-        else:
-            # load images in the directory
-            image_files = glob_images(video_path)
-            image_files.sort()
-            video = []
-            for i in range(len(image_files)):
-                if start_frame is not None and i < start_frame:
-                    continue
-                if end_frame is not None and i >= end_frame:
-                    break
-
-                image_file = image_files[i]
-                image = Image.open(image_file).convert("RGB")
-
-                if bucket_selector is not None and bucket_reso is None:
-                    bucket_reso = bucket_selector.get_bucket_resolution(image.size)  # calc resolution from first frame
-                image = np.array(image)
-                if bucket_reso is not None:
-                    image = resize_image_to_bucket(image, bucket_reso)
-
-                video.append(image)
-    else:
-        # drop frames to match the target fps TODO commonize this code with the above if this works
-        frame_index_delta = target_fps / source_fps  # example: 16 / 30 = 0.5333
-        if os.path.isfile(video_path):
-            container = av.open(video_path)
-            video = []
-            frame_index_with_fraction = 0.0
-            previous_frame_index = -1
-            for i, frame in enumerate(container.decode(video=0)):
-                target_frame_index = int(frame_index_with_fraction)
-                frame_index_with_fraction += frame_index_delta
-
-                if target_frame_index == previous_frame_index:  # drop this frame
-                    continue
-
-                # accept this frame
-                previous_frame_index = target_frame_index
-
-                if start_frame is not None and target_frame_index < start_frame:
-                    continue
-                if end_frame is not None and target_frame_index >= end_frame:
-                    break
-                frame = frame.to_image()
-
-                if bucket_selector is not None and bucket_reso is None:
-                    bucket_reso = bucket_selector.get_bucket_resolution(frame.size)  # calc resolution from first frame
-
-                if bucket_reso is not None:
-                    frame = resize_image_to_bucket(frame, bucket_reso)
-                else:
-                    frame = np.array(frame)
-
-                video.append(frame)
-            container.close()
-        else:
-            # load images in the directory
-            image_files = glob_images(video_path)
-            image_files.sort()
-            video = []
-            frame_index_with_fraction = 0.0
-            previous_frame_index = -1
-            for i in range(len(image_files)):
-                target_frame_index = int(frame_index_with_fraction)
-                frame_index_with_fraction += frame_index_delta
-
-                if target_frame_index == previous_frame_index:  # drop this frame
-                    continue
-
-                # accept this frame
-                previous_frame_index = target_frame_index
-
-                if start_frame is not None and target_frame_index < start_frame:
-                    continue
-                if end_frame is not None and target_frame_index >= end_frame:
-                    break
-
-                image_file = image_files[i]
-                image = Image.open(image_file).convert("RGB")
-
-                if bucket_selector is not None and bucket_reso is None:
-                    bucket_reso = bucket_selector.get_bucket_resolution(image.size)  # calc resolution from first frame
-                image = np.array(image)
-                if bucket_reso is not None:
-                    image = resize_image_to_bucket(image, bucket_reso)
-
-                video.append(image)
-
-    return video
-
-
-class BucketBatchManager:
-    def __init__(
-        self, bucketed_item_info: dict[tuple[Any], list[ItemInfo]], batch_size: int, num_timestep_buckets: Optional[int] = None
-    ):
-        self.batch_size = batch_size
-        self.buckets = bucketed_item_info
-        self.bucket_resos = list(self.buckets.keys())
-        self.bucket_resos.sort()
-        self.num_timestep_buckets = num_timestep_buckets
-        self.timestep_pool = None
-
-        # indices for enumerating batches. each batch is reso + batch_idx. reso is (width, height) or (width, height, frames)
-        self.bucket_batch_indices: list[tuple[tuple[Any], int]] = []
-        for bucket_reso in self.bucket_resos:
-            bucket = self.buckets[bucket_reso]
-            num_batches = math.ceil(len(bucket) / self.batch_size)
-            for i in range(num_batches):
-                self.bucket_batch_indices.append((bucket_reso, i))
-
-        # do no shuffle here to avoid multiple datasets have different order
-        # self.shuffle()
-
-    def show_bucket_info(self):
-        for bucket_reso in self.bucket_resos:
-            bucket = self.buckets[bucket_reso]
-            logger.info(f"bucket: {bucket_reso}, count: {len(bucket)}")
-
-        logger.info(f"total batches: {len(self)}")
-
-    def shuffle(self):
-        # shuffle each bucket
-        for bucket in self.buckets.values():
-            random.shuffle(bucket)
-
-        # shuffle the order of batches
-        random.shuffle(self.bucket_batch_indices)
-
-        if self.num_timestep_buckets is not None and self.num_timestep_buckets > 1:
-            # prepare timesteps for each timestep buckets
-
-            # 1. Calculate total number of timesteps needed for the entire epoch
-            num_batches = len(self.bucket_batch_indices)
-            total_timesteps_needed = num_batches * self.batch_size
-
-            # 2. Generate a single large pool of stratified timesteps
-            all_timesteps = []
-            samples_per_bucket = math.ceil(total_timesteps_needed / self.num_timestep_buckets)
-
-            for i in range(self.num_timestep_buckets):
-                min_t = i / self.num_timestep_buckets
-                max_t = (i + 1) / self.num_timestep_buckets
-                for _ in range(samples_per_bucket):
-                    all_timesteps.append(random.uniform(min_t, max_t))
-
-            # 3. Shuffle the entire pool thoroughly
-            random.shuffle(all_timesteps)
-
-            # Trim the excess timesteps to match the exact number needed
-            all_timesteps = all_timesteps[:total_timesteps_needed]
-
-            # 4. Create the final timestep pool by chunking the shuffled list
-            self.timestep_pool = []
-            for i in range(num_batches):
-                start_idx = i * self.batch_size
-                end_idx = start_idx + self.batch_size
-                self.timestep_pool.append(all_timesteps[start_idx:end_idx])
-                # print(f"timestep pool {i}: {self.timestep_pool[-1]}")
-
-    def __len__(self):
-        return len(self.bucket_batch_indices)
-
-    def __getitem__(self, idx):
-        bucket_reso, batch_idx = self.bucket_batch_indices[idx]
-        bucket = self.buckets[bucket_reso]
-        start = batch_idx * self.batch_size
-        end = min(start + self.batch_size, len(bucket))
-
-        batch_tensor_data = {}
-        varlen_keys = set()
-        for item_info in bucket[start:end]:
-            sd_latent = load_file(item_info.latent_cache_path)
-            sd_te = load_file(item_info.text_encoder_output_cache_path)
-            sd = {**sd_latent, **sd_te}
-
-            # TODO refactor this
-            for key in sd.keys():
-                is_varlen_key = key.startswith("varlen_")  # varlen keys are not stacked
-                content_key = key
-
-                if is_varlen_key:
-                    content_key = content_key.replace("varlen_", "")
-
-                if content_key.endswith("_mask"):
-                    pass
-                else:
-                    content_key = content_key.rsplit("_", 1)[0]  # remove dtype
-                    if content_key.startswith("latents_"):
-                        content_key = content_key.rsplit("_", 1)[0]  # remove FxHxW
-
-                if content_key not in batch_tensor_data:
-                    batch_tensor_data[content_key] = []
-                batch_tensor_data[content_key].append(sd[key])
-
-                if is_varlen_key:
-                    varlen_keys.add(content_key)
-
-        for key in batch_tensor_data.keys():
-            if key not in varlen_keys:
-                batch_tensor_data[key] = torch.stack(batch_tensor_data[key])
-
-        if self.timestep_pool is not None:
-            batch_tensor_data["timesteps"] = self.timestep_pool[idx][: end - start]  # use the pre-generated timesteps
-        else:
-            batch_tensor_data["timesteps"] = None
-
-        return batch_tensor_data
-
-
-class ContentDatasource:
-    def __init__(self):
-        self.caption_only = False  # set to True to only fetch caption for Text Encoder caching
-        self.has_control = False
-
-    def set_caption_only(self, caption_only: bool):
-        self.caption_only = caption_only
-
-    def is_indexable(self):
-        return False
-
-    def get_caption(self, idx: int) -> tuple[str, str]:
-        """
-        Returns caption. May not be called if is_indexable() returns False.
-        """
-        raise NotImplementedError
-
-    def __len__(self):
-        raise NotImplementedError
-
-    def __iter__(self):
-        raise NotImplementedError
-
-    def __next__(self):
-        raise NotImplementedError
-
-
-class ImageDatasource(ContentDatasource):
-    def __init__(self):
-        super().__init__()
-
-    def get_image_data(self, idx: int) -> tuple[str, list[Image.Image], str, list[Image.Image]]:
-        """
-        Returns image data as a tuple of image path, image, and caption for the given index.
-        Key must be unique and valid as a file name.
-        May not be called if is_indexable() returns False.
-        """
-        raise NotImplementedError
-
-
-class ImageDirectoryDatasource(ImageDatasource):
-    def __init__(
-        self,
-        image_directory: str,
-        caption_extension: Optional[str] = None,
-        control_directory: Optional[str] = None,
-        control_count_per_image: Optional[int] = None,
-        multiple_target: bool = False,
-    ):
-        super().__init__()
-        self.image_directory = image_directory
-        self.caption_extension = caption_extension
-        self.control_directory = control_directory
-        self.control_count_per_image = control_count_per_image
-        self.multiple_target = multiple_target
-        self.current_idx = 0
-
-        # glob images
-        logger.info(f"glob images in {self.image_directory}")
-        self.image_paths = glob_images(self.image_directory, caption_extension=self.caption_extension)
-        logger.info(f"found {len(self.image_paths)} images")
-
-        # check if multiple-target images exist
-        self.target_paths: dict[str, list[str]] = {}  # image_path -> list of target image paths
-
-        if self.multiple_target:
-            # sort by length, longer first
-            sorted_image_paths = sorted(self.image_paths, key=lambda p: len(os.path.basename(p)), reverse=True)
-
-            all_image_paths = set(glob_images(self.image_directory))  # image1.jpg, image1_1.jpg, image1_2.jpg, ...
-            multiple_target_candidates = all_image_paths - set(sorted_image_paths)  # those not in the images with captions
-
-            if len(multiple_target_candidates) > 0:
-                logger.info("checking for multiple-target images")
-                for image_path in sorted_image_paths:
-                    image_path_no_ext = os.path.splitext(image_path)[0]
-
-                    # find matching multiple-target images
-                    potential_paths = [p for p in multiple_target_candidates if p.startswith(image_path_no_ext + "_")]
-
-                    if potential_paths:
-                        # sort by the digits (`_0000`) suffix
-                        def sort_key(path):
-                            path_no_ext = os.path.splitext(path)[0]
-                            digits_suffix = path_no_ext.rsplit("_", 1)[-1]
-                            if not digits_suffix.isdigit():
-                                raise ValueError(
-                                    f"Invalid digits suffix in '{path_no_ext}'. Expected a numeric suffix after '_' "
-                                    f"(e.g., '_0', '_1', '_2') for proper sorting of multiple target images."
-                                )
-                            return int(digits_suffix)
-
-                        potential_paths.sort(key=sort_key)
-                        self.target_paths[image_path] = potential_paths
-
-                        # remove to avoid duplicate matching
-                        multiple_target_candidates.difference_update(potential_paths)
-
-                # check the number of targets: all multiple-target images should have the same number of targets
-                num_targets = 0
-                for image_path, paths in self.target_paths.items():
-                    if num_targets == 0:
-                        num_targets = len(paths)
-                    elif num_targets != len(paths):
-                        logger.error(
-                            f"All multiple-target images must have the same number of targets / 全ての複数ターゲット画像は同じ数のターゲットを持つ必要があります: {image_path}"
-                        )
-                        raise ValueError(
-                            f"All multiple-target images must have the same number of targets / 全ての複数ターゲット画像は同じ数のターゲットを持つ必要があります: {image_path}"
-                        )
-
-                if num_targets == 0:
-                    logger.error("no multiple-target images found, but multiple_target is set to True")
-                    raise ValueError("no multiple-target images found, but multiple_target is set to True")
-
-                logger.info(f"found multiple-target images, max targets per image: {num_targets}")
-
-        # glob control images if specified
-        if self.control_directory is not None:
-            logger.info(f"glob control images in {self.control_directory}")
-            self.has_control = True
-            self.control_paths = {}
-
-            # sort image paths for matching control images properly: longer names first
-            image_paths_sorted = sorted(self.image_paths, key=lambda p: len(os.path.basename(p)), reverse=True)
-
-            # glob control images first
-            all_control_image_paths = set(glob_images(self.control_directory))
-
-            for image_path in image_paths_sorted:
-                image_basename = os.path.basename(image_path)
-                image_basename_no_ext = os.path.splitext(image_basename)[0]
-
-                # find matching control images
-                potential_paths = [
-                    p
-                    for p in all_control_image_paths
-                    if os.path.basename(p).startswith(image_basename_no_ext + ".")
-                    or os.path.basename(p).startswith(image_basename_no_ext + "_")
-                ]
-
-                # remove to avoid duplicate matching
-                all_control_image_paths.difference_update(potential_paths)
-
-                if potential_paths:
-                    # sort by the digits (`_0000`) suffix, prefer the one without the suffix
-                    def sort_key(path):
-                        basename = os.path.basename(path)
-                        basename_no_ext = os.path.splitext(basename)[0]
-                        if image_basename_no_ext == basename_no_ext:  # prefer the one without suffix
-                            return 0
-                        digits_suffix = basename_no_ext.rsplit("_", 1)[-1]
-                        if not digits_suffix.isdigit():
-                            raise ValueError(f"Invalid digits suffix in {basename_no_ext}")
-                        return int(digits_suffix) + 1
-
-                    potential_paths.sort(key=sort_key)
-                    if control_count_per_image is not None and len(potential_paths) < control_count_per_image:
-                        logger.error(
-                            f"Not enough control images for {image_path}: found {len(potential_paths)}, expected {control_count_per_image}"
-                        )
-                        raise ValueError(
-                            f"Not enough control images for {image_path}: found {len(potential_paths)}, expected {control_count_per_image}"
-                        )
-
-                    # take the first `control_count_per_image` paths
-                    self.control_paths[image_path] = (
-                        potential_paths[:control_count_per_image] if control_count_per_image is not None else potential_paths
-                    )
-            logger.info(
-                f"found {len(self.control_paths)} matching control images for {'arbitrary' if control_count_per_image is None else control_count_per_image} images"
-            )
-
-            # log the distribution of number of control images
-            count_of_num_control_images = {}
-            for paths in self.control_paths.values():
-                count = len(paths)
-                if count not in count_of_num_control_images:
-                    count_of_num_control_images[count] = 0
-                count_of_num_control_images[count] += 1
-            for count, num_images in count_of_num_control_images.items():
-                logger.info(f"  {num_images} images have {count} control images")
-
-            missing_controls = len(self.image_paths) - len(self.control_paths)
-            if missing_controls > 0:
-                missing_control_paths = set(self.image_paths) - set(self.control_paths.keys())
-                logger.error(f"Could not find matching control images for {missing_controls} images: {missing_control_paths}")
-                raise ValueError(f"Could not find matching control images for {missing_controls} images")
-
-    def is_indexable(self):
-        return True
-
-    def __len__(self):
-        return len(self.image_paths)
-
-    def get_image_data(self, idx: int) -> tuple[str, list[Image.Image], str, Optional[list[Image.Image]]]:
-        image_path = self.image_paths[idx]
-        image_paths = [image_path]
-        if self.multiple_target:
-            # load multiple-target images
-            image_paths += self.target_paths.get(image_path, [])
-
-        images = []
-        for p in image_paths:
-            img = Image.open(p)
-            if img.mode != "RGB" and img.mode != "RGBA":
-                img = img.convert("RGB")
-            images.append(img)
-
-        _, caption = self.get_caption(idx)
-
-        controls = None
-        if self.has_control:
-            controls = []
-            for control_path in self.control_paths[image_path]:
-                control = Image.open(control_path)
-                if control.mode != "RGB" and control.mode != "RGBA":
-                    control = control.convert("RGB")
-                controls.append(control)
-
-        return image_path, images, caption, controls
-
-    def get_caption(self, idx: int) -> tuple[str, str]:
-        image_path = self.image_paths[idx]
-        caption_path = os.path.splitext(image_path)[0] + self.caption_extension if self.caption_extension else ""
-        with open(caption_path, "r", encoding="utf-8") as f:
-            caption = f.read().strip()
-        return image_path, caption
-
-    def __iter__(self):
-        self.current_idx = 0
-        return self
-
-    def __next__(self) -> callable:
-        """
-        Returns a fetcher function that returns image data.
-        """
-        if self.current_idx >= len(self.image_paths):
-            raise StopIteration
-
-        if self.caption_only:
-
-            def create_caption_fetcher(index):
-                return lambda: self.get_caption(index)
-
-            fetcher = create_caption_fetcher(self.current_idx)
-        else:
-
-            def create_image_fetcher(index):
-                return lambda: self.get_image_data(index)
-
-            fetcher = create_image_fetcher(self.current_idx)
-
-        self.current_idx += 1
-        return fetcher
-
-
-class ImageJsonlDatasource(ImageDatasource):
-    def __init__(self, image_jsonl_file: str, control_count_per_image: Optional[int] = None, multiple_target: bool = False):
-        super().__init__()
-        self.image_jsonl_file = image_jsonl_file
-        self.control_count_per_image = control_count_per_image
-        self.multiple_target = multiple_target
-        self.current_idx = 0
-
-        # load jsonl
-        logger.info(f"load image jsonl from {self.image_jsonl_file}")
-        self.data = []
-        with open(self.image_jsonl_file, "r", encoding="utf-8") as f:
-            for line in f:
-                try:
-                    data = json.loads(line)
-                except json.JSONDecodeError:
-                    logger.error(f"failed to load json: {line} @ {self.image_jsonl_file}")
-                    raise
-                self.data.append(data)
-        logger.info(f"loaded {len(self.data)} images")
-
-        # Normalize control paths
-        for item in self.data:
-            if "control_path" in item:
-                item["control_path_0"] = item.pop("control_path")
-
-            # Ensure control paths are named consistently, from control_path_0000 to control_path_0, control_path_1, etc.
-            control_path_keys = [key for key in item.keys() if key.startswith("control_path_")]
-            control_path_keys.sort(key=lambda x: int(x.split("_")[-1]))
-            for i, key in enumerate(control_path_keys):
-                if key != f"control_path_{i}":
-                    item[f"control_path_{i}"] = item.pop(key)
-
-        # Check if there are control paths in the JSONL
-        self.has_control = any("control_path_0" in item for item in self.data)
-        if self.has_control:
-            if self.control_count_per_image is None:
-                logger.info(f"found {len(self.data)} images with arbitrary control images per image in JSONL data")
-            else:
-                missing_control_images = [
-                    item["image_path"]
-                    for item in self.data
-                    if sum(f"control_path_{i}" not in item for i in range(self.control_count_per_image)) > 0
-                ]
-                if missing_control_images:
-                    logger.error(f"Some images do not have control paths in JSONL data: {missing_control_images}")
-                    raise ValueError(f"Some images do not have control paths in JSONL data: {missing_control_images}")
-                logger.info(
-                    f"found {len(self.data)} images with {self.control_count_per_image} control images per image in JSONL data"
-                )
-
-    def is_indexable(self):
-        return True
-
-    def __len__(self):
-        return len(self.data)
-
-    def get_image_data(self, idx: int) -> tuple[str, list[Image.Image], str, Optional[list[Image.Image]]]:
-        data = self.data[idx]
-        image_path = data.get("image_path", data.get("image_path_0"))
-        image_paths = [image_path]
-        if self.multiple_target:
-            # load multiple-target images
-            while True:
-                next_index = len(image_paths)  # start from 1
-                next_image_path = data.get("image_path_" + str(next_index), None)
-                if next_image_path is None:
-                    break
-                if not os.path.exists(next_image_path):
-                    raise ValueError(f"multiple-target image not found: {next_image_path}")
-
-                image_paths.append(next_image_path)
-
-        images = []
-        for path in image_paths:
-            img = Image.open(path)
-            if img.mode != "RGB" and img.mode != "RGBA":
-                img = img.convert("RGB")
-            images.append(img)
-
-        caption = data["caption"]
-
-        controls = None
-        if self.has_control:
-            controls = []
-            for i in range(self.control_count_per_image or 1000):  # arbitrary large number if control_count_per_image is None
-                if f"control_path_{i}" not in data:
-                    break
-                control_path = data[f"control_path_{i}"]
-                control = Image.open(control_path)
-                if control.mode != "RGB" and control.mode != "RGBA":
-                    control = control.convert("RGB")
-                controls.append(control)
-
-        return image_path, images, caption, controls
-
-    def get_caption(self, idx: int) -> tuple[str, str]:
-        data = self.data[idx]
-        image_path = data.get("image_path", data.get("image_path_0"))
-        caption = data["caption"]
-        return image_path, caption
-
-    def __iter__(self):
-        self.current_idx = 0
-        return self
-
-    def __next__(self) -> callable:
-        if self.current_idx >= len(self.data):
-            raise StopIteration
-
-        if self.caption_only:
-
-            def create_caption_fetcher(index):
-                return lambda: self.get_caption(index)
-
-            fetcher = create_caption_fetcher(self.current_idx)
-
-        else:
-
-            def create_fetcher(index):
-                return lambda: self.get_image_data(index)
-
-            fetcher = create_fetcher(self.current_idx)
-
-        self.current_idx += 1
-        return fetcher
-
-
-class VideoDatasource(ContentDatasource):
-    def __init__(self):
-        super().__init__()
-
-        # None means all frames
-        self.start_frame = None
-        self.end_frame = None
-
-        self.bucket_selector = None
-
-        self.source_fps = None
-        self.target_fps = None
-
-    def __len__(self):
-        raise NotImplementedError
-
-    def get_video_data_from_path(
-        self,
-        video_path: str,
-        start_frame: Optional[int] = None,
-        end_frame: Optional[int] = None,
-        bucket_selector: Optional[BucketSelector] = None,
-    ) -> list[Image.Image]:
-        # this method can resize the video if bucket_selector is given to reduce the memory usage
-
-        start_frame = start_frame if start_frame is not None else self.start_frame
-        end_frame = end_frame if end_frame is not None else self.end_frame
-        bucket_selector = bucket_selector if bucket_selector is not None else self.bucket_selector
-
-        video = load_video(
-            video_path, start_frame, end_frame, bucket_selector, source_fps=self.source_fps, target_fps=self.target_fps
-        )
-        return video
-
-    def get_control_data_from_path(
-        self,
-        control_path: str,
-        start_frame: Optional[int] = None,
-        end_frame: Optional[int] = None,
-        bucket_selector: Optional[BucketSelector] = None,
-    ) -> list[Image.Image]:
-        start_frame = start_frame if start_frame is not None else self.start_frame
-        end_frame = end_frame if end_frame is not None else self.end_frame
-        bucket_selector = bucket_selector if bucket_selector is not None else self.bucket_selector
-
-        control = load_video(
-            control_path, start_frame, end_frame, bucket_selector, source_fps=self.source_fps, target_fps=self.target_fps
-        )
-        return control
-
-    def set_start_and_end_frame(self, start_frame: Optional[int], end_frame: Optional[int]):
-        self.start_frame = start_frame
-        self.end_frame = end_frame
-
-    def set_bucket_selector(self, bucket_selector: BucketSelector):
-        self.bucket_selector = bucket_selector
-
-    def set_source_and_target_fps(self, source_fps: Optional[float], target_fps: Optional[float]):
-        self.source_fps = source_fps
-        self.target_fps = target_fps
-
-    def __iter__(self):
-        raise NotImplementedError
-
-    def __next__(self):
-        raise NotImplementedError
-
-
-class VideoDirectoryDatasource(VideoDatasource):
-    def __init__(self, video_directory: str, caption_extension: Optional[str] = None, control_directory: Optional[str] = None):
-        super().__init__()
-        self.video_directory = video_directory
-        self.caption_extension = caption_extension
-        self.control_directory = control_directory  # 新しく追加: コントロール画像ディレクトリ
-        self.current_idx = 0
-
-        # glob videos
-        logger.info(f"glob videos in {self.video_directory}")
-        self.video_paths = glob_videos(self.video_directory)
-        logger.info(f"found {len(self.video_paths)} videos")
-
-        # glob control images if specified
-        if self.control_directory is not None:
-            logger.info(f"glob control videos in {self.control_directory}")
-            self.has_control = True
-            self.control_paths = {}
-            for video_path in self.video_paths:
-                video_basename = os.path.basename(video_path)
-                # construct control path from video path
-                # for example: video_path = "vid/video.mp4" -> control_path = "control/video.mp4"
-                control_path = os.path.join(self.control_directory, video_basename)
-                if os.path.exists(control_path):
-                    self.control_paths[video_path] = control_path
-                else:
-                    # use the same base name for control path
-                    base_name = os.path.splitext(video_basename)[0]
-
-                    # directory with images. for example: video_path = "vid/video.mp4" -> control_path = "control/video"
-                    potential_path = os.path.join(self.control_directory, base_name)  # no extension
-                    if os.path.isdir(potential_path):
-                        self.control_paths[video_path] = potential_path
-                    else:
-                        # another extension for control path
-                        # for example: video_path = "vid/video.mp4" -> control_path = "control/video.mov"
-                        for ext in VIDEO_EXTENSIONS:
-                            potential_path = os.path.join(self.control_directory, base_name + ext)
-                            if os.path.exists(potential_path):
-                                self.control_paths[video_path] = potential_path
-                                break
-
-            logger.info(f"found {len(self.control_paths)} matching control videos/images")
-            # check if all videos have matching control paths, if not, raise an error
-            missing_controls = len(self.video_paths) - len(self.control_paths)
-            if missing_controls > 0:
-                # logger.warning(f"Could not find matching control videos/images for {missing_controls} videos")
-                missing_controls_videos = [video_path for video_path in self.video_paths if video_path not in self.control_paths]
-                logger.error(
-                    f"Could not find matching control videos/images for {missing_controls} videos: {missing_controls_videos}"
-                )
-                raise ValueError(f"Could not find matching control videos/images for {missing_controls} videos")
-
-    def is_indexable(self):
-        return True
-
-    def __len__(self):
-        return len(self.video_paths)
-
-    def get_video_data(
-        self,
-        idx: int,
-        start_frame: Optional[int] = None,
-        end_frame: Optional[int] = None,
-        bucket_selector: Optional[BucketSelector] = None,
-    ) -> tuple[str, list[Image.Image], str, Optional[list[Image.Image]]]:
-        video_path = self.video_paths[idx]
-        video = self.get_video_data_from_path(video_path, start_frame, end_frame, bucket_selector)
-
-        _, caption = self.get_caption(idx)
-
-        control = None
-        if self.control_directory is not None and video_path in self.control_paths:
-            control_path = self.control_paths[video_path]
-            control = self.get_control_data_from_path(control_path, start_frame, end_frame, bucket_selector)
-
-        return video_path, video, caption, control
-
-    def get_caption(self, idx: int) -> tuple[str, str]:
-        video_path = self.video_paths[idx]
-        caption_path = os.path.splitext(video_path)[0] + self.caption_extension if self.caption_extension else ""
-        with open(caption_path, "r", encoding="utf-8") as f:
-            caption = f.read().strip()
-        return video_path, caption
-
-    def __iter__(self):
-        self.current_idx = 0
-        return self
-
-    def __next__(self):
-        if self.current_idx >= len(self.video_paths):
-            raise StopIteration
-
-        if self.caption_only:
-
-            def create_caption_fetcher(index):
-                return lambda: self.get_caption(index)
-
-            fetcher = create_caption_fetcher(self.current_idx)
-
-        else:
-
-            def create_fetcher(index):
-                return lambda: self.get_video_data(index)
-
-            fetcher = create_fetcher(self.current_idx)
-
-        self.current_idx += 1
-        return fetcher
-
-
-class VideoJsonlDatasource(VideoDatasource):
-    def __init__(self, video_jsonl_file: str):
-        super().__init__()
-        self.video_jsonl_file = video_jsonl_file
-        self.current_idx = 0
-
-        # load jsonl
-        logger.info(f"load video jsonl from {self.video_jsonl_file}")
-        self.data = []
-        with open(self.video_jsonl_file, "r", encoding="utf-8") as f:
-            for line in f:
-                data = json.loads(line)
-                self.data.append(data)
-        logger.info(f"loaded {len(self.data)} videos")
-
-        # Check if there are control paths in the JSONL
-        self.has_control = any("control_path" in item for item in self.data)
-        if self.has_control:
-            control_count = sum(1 for item in self.data if "control_path" in item)
-            if control_count < len(self.data):
-                missing_control_videos = [item["video_path"] for item in self.data if "control_path" not in item]
-                logger.error(f"Some videos do not have control paths in JSONL data: {missing_control_videos}")
-                raise ValueError(f"Some videos do not have control paths in JSONL data: {missing_control_videos}")
-            logger.info(f"found {control_count} control videos/images in JSONL data")
-
-    def is_indexable(self):
-        return True
-
-    def __len__(self):
-        return len(self.data)
-
-    def get_video_data(
-        self,
-        idx: int,
-        start_frame: Optional[int] = None,
-        end_frame: Optional[int] = None,
-        bucket_selector: Optional[BucketSelector] = None,
-    ) -> tuple[str, list[Image.Image], str, Optional[list[Image.Image]]]:
-        data = self.data[idx]
-        video_path = data["video_path"]
-        video = self.get_video_data_from_path(video_path, start_frame, end_frame, bucket_selector)
-
-        caption = data["caption"]
-
-        control = None
-        if "control_path" in data and data["control_path"]:
-            control_path = data["control_path"]
-            control = self.get_control_data_from_path(control_path, start_frame, end_frame, bucket_selector)
-
-        return video_path, video, caption, control
-
-    def get_caption(self, idx: int) -> tuple[str, str]:
-        data = self.data[idx]
-        video_path = data["video_path"]
-        caption = data["caption"]
-        return video_path, caption
-
-    def __iter__(self):
-        self.current_idx = 0
-        return self
-
-    def __next__(self):
-        if self.current_idx >= len(self.data):
-            raise StopIteration
-
-        if self.caption_only:
-
-            def create_caption_fetcher(index):
-                return lambda: self.get_caption(index)
-
-            fetcher = create_caption_fetcher(self.current_idx)
-
-        else:
-
-            def create_fetcher(index):
-                return lambda: self.get_video_data(index)
-
-            fetcher = create_fetcher(self.current_idx)
-
-        self.current_idx += 1
-        return fetcher
+# The following classes have been moved to datasources.py but are kept here
+# as a comment reference. They are re-imported above for backward compatibility.
+# - ContentDatasource, ImageDatasource, ImageDirectoryDatasource, ImageJsonlDatasource
+# - VideoDatasource, VideoDirectoryDatasource, VideoJsonlDatasource
 
 
 class BaseDataset(torch.utils.data.Dataset):
